Add files via upload
homeF basically stands for the homefeed page which i assume u guys were already working on, you might have to take what i did in the homeF page and implement that into our homepage since i saw a part were the feed is supposed to go on there. the homeFitem is the actually feedPost itself, with text user id and (a very ugly) time stamp showing when it was posted. there are two messages.js and two users.js so basically theres a message.js and a user.js for the /action -> constants, and the reducers